+++ internal/action/release_triggered_event_handler.go
@@ -1,4 +1,4 @@
-package deployment
+package action
 
 import (
 	"context"
